// src/app/shared/types/api-output-types/airworthiness-management/generate-airworthiness-control-table-output.interface.ts
import { BidoEquipmentDTO } from '../../api-types/bido-equipment-dto.interface';

export interface GenerateAirworthinessControlTableOutput {
  bidoEquipment: BidoEquipmentDTO;
  children: GenerateAirworthinessControlTableOutput[];
  assetRepresentation?: string;
  adsbAlertIcon?: string;
  adsbAlertTooltip?: string;
  airworthinessAlertIcon?: string;
  airworthinessAlertTooltip?: string;
  assetDesignation?: string;
  checkResultIcon?: string;
  checkResultReport?: string;
  checkResultTooltip?: string;
  closestCalendarLimitAsset?: string;
  closestCalendarLimitDate?: Date;
  code?: string;
  counterAlertStatus?: string;
  counterAlertTooltip?: string;
  counterCsn?: string;
  counterCsnIcon?: string;
  counterCsnTooltip?: string;
  counterTsn?: string;
  counterTsnIcon?: string;
  counterTsnTooltip?: string;
  docValidityEndDate?: string;
  docValidityEndDateIcon?: string;
  docValidityEndDateTooltip?: string;
  globalAirworthinessStatus?: string;
  maintenanceProgram?: string;
  maintenanceProgramStatus?: string;
  maintenanceProgramTooltip?: string;
  nextWorkPackage?: string;
  nextWorkPackageAlert?: boolean;
  nextWorkPackageTooltip?: string;
  potential?: string;
  potentialIcon?: string;
  smallestRemainingInCyclesAsset?: string;
  smallestRemainingInCyclesValue?: number;
  smallestRemainingInHoursAsset?: string;
  smallestRemainingInHoursValue?: number;
  smallestRemainingLifeInCycles?: string;
  smallestRemainingLifeInDays?: string;
  smallestRemainingLifeInHours?: string;
  status?: string;
  remaining?: string;
  remainingCycles?: string;
  remainingDays?: string;
  remainingEquivalent?: string;
  remainingEquivalentIcon?: string;
  remainingEquivalentTooltip?: string;
  remainingEquivalentValue?: number;
  remainingHours?: string;
  subElementQuantity?: number;
}